# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Data Analyst API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /datasets/{id}:
    get:
      operationId: getDataset
      tags: [datasets]
      summary: Get dataset metadata and overview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dataset details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DatasetDetail"
        "404":
          description: Not found
    delete:
      operationId: deleteDataset
      tags: [datasets]
      summary: Delete a dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
...
  /datasets/{id}/report:
    post:
      operationId: generateReport
      tags: [reports]
      summary: Generate a PDF report for the dataset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportInput"
      responses:
        "200":
          description: Report generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportResult"
...
components:
  schemas:
...
    DatasetDetail:
      type: object
      required: [id, name, filename, rowCount, columnCount, uploadedAt, fileSize, columns]
      properties:
        id:
          type: string
        name:
          type: string
        filename:
          type: string
        rowCount:
          type: integer
        columnCount:
          type: integer
        uploadedAt:
          type: string
        fileSize:
          type: integer
        columns:
          type: array
          items:
            $ref: "#/components/schemas/ColumnInfo"
        previewRows:
          type: array
          items:
            type: object
    ColumnInfo:
      type: object
      required: [name, dtype, nullCount, uniqueCount]
      properties:
        name:
          type: string
        dtype:
          type: string
        nullCount:
          type: integer
        uniqueCount:
          type: integer
        min:
          type: ["string", "null"]
        max:
          type: ["string", "null"]
        mean:
          type: ["number", "null"]
        std:
          type: ["number", "null"]
...
    ReportInput:
      type: object
      properties:
        title:
          type: string
        includeCharts:
          type: boolean
        includeInsights:
          type: boolean
        includeKpis:
          type: boolean
    ReportResult:
      type: object
      required: [id, datasetId, title, createdAt, downloadUrl]
      properties:
        id:
          type: string
        datasetId:
          type: string
        title:
          type: string
        createdAt:
          type: string
        downloadUrl:
          type: string
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]