# lib/api-zod/src/generated/types/datasetDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */
import type { ColumnInfo } from './columnInfo';
import type { DatasetDetailPreviewRowsItem } from './datasetDetailPreviewRowsItem';

export interface DatasetDetail {
  id: string;
  name: string;
  filename: string;
  rowCount: number;
  columnCount: number;
  uploadedAt: string;
  fileSize: number;
  columns: ColumnInfo[];
  previewRows?: DatasetDetailPreviewRowsItem[];
}