# lib/api-zod/src/generated/types/reportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportInput {
  title?: string;
  includeCharts?: boolean;
  includeInsights?: boolean;
  includeKpis?: boolean;
}