# lib/api-zod/src/generated/types/reportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Data Analyst API
 * OpenAPI spec version: 0.1.0
 */

export interface ReportResult {
  id: string;
  datasetId: string;
  title: string;
  createdAt: string;
  downloadUrl: string;
}